Izmena tipova podataka u klasi CacheSettings: ushort -> uint

+++ GitHub_API/CacheSettings.cs
@@ -2,9 +2,9 @@
 
 public static class CacheSettings
 {
-    public static ushort MaxEntries { get; private set; } = 1000;
+    public static uint MaxEntries { get; private set; } = 1000;
     public static TimeSpan CleanupPeriod { get; private set; } = TimeSpan.FromMinutes(30);
-    public static ushort MaxEntryContributorCount { get; private set; } = 1000;
+    public static uint MaxEntryContributorCount { get; private set; } = 1000;
     public static bool CachingEnabled { get; private set; } = true;
 
 }
